#34 Extracting JSON/DOT serialization logic from PackageDependencies class
