+++ javac-plugin/src/main/java/org/moditect/deptective/internal/export/ModelSerializer.java
@@ -1,0 +1,33 @@
+/**
+ *  Copyright 2019 The ModiTect authors
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
+package org.moditect.deptective.internal.export;
+
+import org.moditect.deptective.internal.model.Package;
+import org.moditect.deptective.internal.model.WhitelistedPackagePattern;
+
+/**
+ * Implementations serialize a given architecture model into external representations such as JSON or GraphViz.
+ *
+ * @author Gunnar Morling
+ */
+public interface ModelSerializer {
+
+    void addPackage(Package pakkage);
+
+    void addWhitelistedPackagePattern(WhitelistedPackagePattern pattern);
+
+    String serialize();
+}
